+++ Source/DownloadAPIRequest.swift
@@ -71,7 +71,7 @@ open class DownloadAPIRequest<Model, ErrorModel: DownloadErrorSerializable>: Bas
         super.init(path: path, tron: tron)
     }
 
-    override func alamofireRequest(from manager: Session) -> Request? {
+    override func alamofireRequest(from manager: Session) -> Request {
         switch type {
         case .download(let destination):
             return manager.download(urlBuilder.url(forPath: path), method: method, parameters: parameters,
@@ -92,7 +92,7 @@ open class DownloadAPIRequest<Model, ErrorModel: DownloadErrorSerializable>: Bas
      
      - returns: Alamofire.Request or nil if request was stubbed.
      */
-    open func performCollectingTimeline(withCompletion completion: @escaping ((Alamofire.DownloadResponse<Model>) -> Void)) -> DownloadRequest? {
+    open func performCollectingTimeline(withCompletion completion: @escaping ((Alamofire.DownloadResponse<Model>) -> Void)) -> DownloadRequest {
         return performAlamofireRequest(completion)
     }